requirements.txt
python-telegram-bot>=21.0,<22.0
aiosqlite>=0.19,<1.0
python-dotenv>=1.0,<2.0
